Disable saving coverage report
This doesn't work in Docker on Linux because of differences in uid and fs
permissions and all that. We'd need to rework how everything gets run
and which uid/gid it gets run as to get this to work again. In the meantime,
I'm just going to shut it off.