bug 1552950: Add GCS Emulator service
Add gcs-emulator, an emulator for Google Cloud Storage (GCS), analogous
to minio for AWS S3.

It works like the current GCS integration, allowing uploads of symbols
and zipped symbol packs, and download of symbols via a public host.
The server does not authenticate requests.

The service runs by default, but requires an .env override to enable in
the web and worker services.

A FakeGCSClient, inheriting from the official GCS Python client, is used
to talk to gcs-emulator. This includes using anonymous access, allowing
the self-signed certificate, and changing all the URLs to the local
gcs-emulator URLs. Due to limitations in the GCS Python client, the same
Tecken process can't talk to gcs-emulator and the real GCS service at
the same time.