Upgrade to Python 3.7
This pins to Python 3.7.4 using slim and stretch. We need to keep using
stretch because it has the gcc-6 package which we need to compile
dump_syms. In order to get off of stretch, we'd need to fix all that.
It's probably doable, but I didn't want to tackle it today.