bug 1618942: add make-symbols-zip.py and setup_tests.sh scripts
This adds make-symbols-zip.py which builds symbols ZIP files to use to
test uploading.

This adds a setup_tests.sh script which sets everything up for running
system tests. This also removes a ton of data--we don't need that in the
repository.

This also fixes issues here and there that I found while testing things.