Fix build rule to build frontend and linting images
Previously, the "frontend" image wasn't re-created when you ran "make
build", so if you updated JS dependencies and ran "make build" and
expected that to rebuild everything... SURPRISE!

This also didn't re-create the "linting" image.

Now it does both of these things so testing in a local dev environment
is better now.