bug 1630120: fix frontend in local dev environment
This fixes the frontend container to run with the UID/GID set in the
.env file so it doesn't create files as root.

This adds a script to build the yarn.lock file and dealing with the
frontend/node_modules/ directory that has "stuff" in it and can cause
problems if left around.

This demotes react-scripts back to 3.4.0 which works--3.4.1 seems to
skip running the dev server with no errors.

This rebuilds the yarn.lock file.