bug 1649531: handle empty string debug filename and id
This fixes symbolication to handle empty strings for the debug filename
and debug id in the memoryMap portion of the symbolication request.
Previously, it would raise an error. It no longer raises an error.
Instead, the code doesn't check cache and doesn't try to download the
module and treats it as if the module isn't available.

I also fixed make-stacks so it no longer drops modules where the debug
filename or id aren't known.