bug 1651190: add cleanse_missingsymbol command
This adds a cleanse_missingsymbol command that will delete missingsymbol
records that haven't been modified in the last 30 days. These are
records for symbols where either the symbol is no longer missing or no
one has requested the symbol in 30 days and thus we probably don't care
about it anymore.