bug 1655944: fix middleware to slurp request body before responding
This fixes the token middleware to fully read the request body before
responding with a 403 or similar response. This handles the case where
the client is using chunked transfer encoding, but the entire request
body hasn't been sent, yet. In that scenario, returning a response
causes nginx to drop the connection. This fixes it by reading the entire
body so that the response gets sent.

This adds a system test that tests the stage and prod environments for
the issue.