bug 1636210: lay groundwork for new microservice
This redoes a bunch of project scaffolding to allow for multiple
services in the same project repository.

This also builds the initial project bits of the Eliot service. You can
now run the service and it has Dockerflow endpoints and an index page
and can do 404s and basic stuff. It also has a basic test and test
scaffolding.