Create entrypoint.sh focused on services
This removes the shell implementation of "waitfor" replacing it with
urlwait.

This redoes linting so it uses a bin/run_lint.sh script for python lint
and lintfix and runs in regular or ci containers.

This redoes tests so that it uses a bin/run_test.sh script that sets
things up and uses the test-c or test container depending on whether
it's CI or not.

This adds better auto-generated usage for bin/entrypoint.sh.