Fix systemtests/test_env.sh to only symbolicate 20 stacks
If you have 1000 stacks in the directory, this restricts the script to
testing only 20 of them. This is a quick fix for a problem I had today.
It's not a great fix and we should probably do something different such
that the test is stable run to run over months.