bug 1669248: add dockerhub auth and conform to dockerflow
This ditches the deploy-dockerhub.sh script and adds the bits directly
to the CircleCI config file. It also changes it to conform to the
existing Dockerflow conventions.

This switches to using DOCKERHUB_REPO, DOCKER_USER, and DOCKER_PASS
variables. They'll be in the environment when CircleCI is running a job
based on this repo, but won't be available in the environment when
CircleCI is running a job based on a fork.