bug 1674406: reimplement metrics in Eliot and add new ones
This reimplements how metrics work in Eliot and adds the ones we need
for the bug.

The new system centralizes all the metrics that Eliot generates into a
single file. It adds an extension for extracting that data and
auto-generating documentation. It also shapes the metrics data as a
Markus MetricsFilter so it validates any metrics being emitted.

Adding a new metric involves adding it to ELIOT_METRICS and then adding
the code to emit the metric. That's about how much effort it was before,
so I think this seems promising. I'm inclined to redo Tecken webapp
accordingly.