Fix ResourceWarning errors in tests
Both Tecken and Eliot had ResourceWarning errors when running the tests.

For Tecken, I switched to the pytest tmpdir fixture, cleaned up the code
around that, and then switched from django.views.static.serve to
manually opening the file and serving the contents. The Django 2.2
version of serve opens files and doesn't explicitly close the File
object which causes the ResourceWarning in the test.

For Eliot, the disk cache manager uses inotify and wasn't cleaning up
the io descriptors for it in the tests. This fixes that and tweaks the
code so it's cleaner in regards to which code owns which resources.