Update Django to 2.2.18 and re-generate requirements.txt
This updates Django to 2.2.18 and regenerates the requirements.txt
file picking up other updates.

I think I used an older version of pip-compile and that caused the
dependencies to be listed on a single line.

This regenerates the file and looks like the only change is to
vertically list the dependencies.