bug 1698371: stop-gap fix for ResponseError: WRONGTYPE
Over the weekend, we had something hit the symbolication API and kick up
5,000 of these "ResponseError: WRONGTYPE Operation against a key holding
the wrong kind of value" errors.

We're replacing that code with Eliot at some point. This is a stop-gap
solution so it doesn't page us anymore. I'm setting the length to 0
which puts it in a "the keys exist, but the hashmap doesn't" situation
and I think the code will figure itself out from there.