bug 1699546: return correct value for "module" in symbolication
For Windows SYM files, the debug_filename is something.pdb, but we
really want to return something.dll which is the PE filename. SYM files
encode this in the INFO line.

symbolic parses that line and captures that information, but doesn't
make it available in the symcache, so we had to add some code to peek in
the SYM file and pull out the PE filename if it's there. But then since
we're caching symcache files on disk, we needed to redo disk caching to
cache items that consist of multiple key/val pairs.

This stores the key/val data as a single file using msgpack. That way
the disk cache LRU can think about files as units and we don't have to
deal with the complexities of other possibilities.