bug 1701832: remove MSGPackSerializer
Tecken had a MSGPackSerializer that fixed an issue in django-redis
preventing us from using recent releases of msgpack. django-redis
released a version (4.12.0) that supports recent releases of msgpack, so
we don't need our MSGPackSerializer anymore.

Note that we need to keep msgpack in requirements.in because we use the
MSGPackSerializer from django-redis and django-redis doesn't have
msgpack in an install extras kind of thing.