bug 1731412: force linux/amd64 platform for docker images
This forces the use of linux/amd64 for docker images. People who are
using m1 Macs will now be forced to emulate linux/amd64. I figured I'd
try this out now since it's the easier fix.

Docker has a "known issues" here:

https://docs.docker.com/desktop/mac/apple-silicon/#known-issues

We use inotify in a couple of places. If it turns into a problem, we can
undo this change and try something different