bug 1699776: add exception notification to loop and shutdown
This adds exception notification to the event handling loop. If there
are unhandled exceptions, they'll get sent to Sentry. It's possible the
unhandled exception is not ephemeral, so this keeps track of how many
errors of this kind it's hit and terminates if there are too many. I
don't know if 10 is a magic number, but seems like a good starting point
for a disk cache manager.

This also adds exception handling to shutdown when it's removing inotify
watches.