bug 1731412: downgrade to node 14
This downgrades Tecken to using node 14. Tecken only uses node to build
static assets, so as long as they're building fine, it doesn't really
matter what version of node we're using.

This allows people using macs with m1 processors to build and run
Tecken. Previous to this fix, they would hit this issue:

https://github.com/docker/for-mac/issues/5831