bug 1733636: redo Eliot index page
This adds scaffolding for styling and JS for the Eliot index page. It's
not the most optimized page size-wise, but it requires minimal
maintenance and can just sit around.

We can improve it going forward and we'll probably want to change how
it's built to match Tecken when we overhaul the Tecken frontend ui.