bug 1734610: overhaul symbolication docs
This updates the symbolication API docs to show API changes, the new
host, better examples, and clearer notes about using a user-agent
header.

This also overhauls all the rest of the docs to talk about "Tecken" as
the umbrella project, "Symbols Service" as the symbols service (was
"tecken"), and "Symbolication Service" as the symbolication service (aka
Eliot).