bug 1734641: add debug stats to symbolication v5
This adds support for the Debug header in the symbolication API rewrite
for the v5 version. I predict we'll want to add other stats, so this
implements a flexible infrastructure for adding more stats without
having to pass anything else around.

Note that this doesn't add support for the Debug header for the v4 API
since that's deprecated and going away.