bug 1739202: rework requests wrappers to handle content errors
The session_with_retries() we had would create a requests Session
instance that used an HTTPAdapter that had a default timeout. It would
also create a Retry instance which would retry various error scenarios
of the HTTP request.

However, once the HTTP request is successfully sent, then a Response
object is created and returned. This Response object is now outside the
scope of all the error retry handling.

If the stream of HTTP response payload data pauses and exceeds the read
timeout, then a ReadTimeout error is thrown and the Eliot downloader
treats the attempt to fetch the sym file as an ErrorFileNotFound. It
should instead retry it.

This reworks all the requests scaffolding to keep the timeout code we
had but also cover the case where a network error when reading the HTTP
response payload data also retries the request.