bug 1749309: rework cleanse_upload to skip SELECTs
Because FileUpload has on_delete=SET_NULL, when we delete Upload from
the db, it does a SELECT for all the Upload so it can update all the
FileUpload. That's a ridiculous amount of data and when cleaning our
stage/prod databases, it makes it impossible to run the command.

This (ab)uses a private method _raw_delete() to delete the data without
doing the SELECTs. I think this should work going forward--Django 4.0
still has it.

This also prints the queries that run so we can more easily use/debug
this going forward.