bug 1699776: adjust how disk cache manager is run
This adjusts the run script for the eliot disk cache manager. The
thinking is that in the case the disk cache manager hits an unhandled
exception, the process will send errors to sentry and then die. Without
a running disk cache manager, the node's cache will fill up and the
node's performance will degrade significantly and maybe die.

Maybe the errors are ephemeral? Maybe the errors are recoverable?

This adjusts the run script to run the disk cache manager process in a
loop. If the disk cache manager dies, then the loop will pause for a
minute and start the disk cache manager again. If the error was
ephemeral or recoverable, the disk cache manager will take stock of the
cache, clean it up, and continue functioning. If the error was not
ephemeral or recoverable (a code bug, perhaps), then the disk cache
manager will send errors to sentry, die again, and the loop will pause
for a minute and start it up again.

The loop allows the disk cache manager to recover. The pause reduces
flapping and overwhelming spikes in error reports.

I think this is how I want to run this for now. We can adjust going
forward as we learn more.