bug 1764622: add license_check.py script
This adds a license_check.py script that checks specified files for a
license header. If you pass in a target file or directory, it'll check
that. If you pass in no arguments, it looks at all the files checked
into source control.

If there are files with missing headers, it returns a non-zero exit
code, so it'll fail in CI.