bug 1764622: move licence_check to separate CI job
The license_check.py script should run in the context of a git
repository which we don't have when running it in the tecken docker
image.

Because it doesn't have any dependencies, we can run it in the executor
container and don't need to run it in the tecken docker image. Thus we
set it up as a separate CI job.

This also switches the docker image to cimg/python:3.9.12 because that
has Python in it and cimg/base does not.