bug 1758566: fix cache invalidation for try symbols
When try symbols get uploaded, the "does this exist in s3?" cache wasn't
getting invalidated, so symbol download could continue to kick up 404s.

This fixes that specific issue by adding the try symbols source to the
downloader hard-coded in the upload code. There's another bug to rewrite
how all this works and we can fix it in a more resilient way then.