bug 1771319: fix CI issues
I removed systemtests/ from .dockerignore so the lint job runs; we
should have the lint job working on a checkout rather than the image,
but that's difficult to configure with the way we have things
structured currently so I'm going to think about that some other day.

I fixed circleci to start the localstack service instead of minio which
is gone now.