bug 1764569: implement sentry scrubbing
sentry_sdk doesn't scrub sentry events before sending them to the Sentry
server. This adds libsentry which has a scrubbing system as well as a
set_up_sentry more tailored to our uses which we'll share with other
socorro-eng projects.

This sets up the Boto3, Redis, and Django integrations as well as the
default ones sentry_sdk sets up.

This makes some initial decisions on what to scrub. Generally, it tries
to scrub anything related to auth tokens, the user's session id, and
user oauth state.