bug 1779633: update symbolic and emit inlines
The update to major version 9 required the following changes:

 - `LineInfo` has been replaced with `SourceLocation`, so I renamed the
    `lineinfo` variable to `sourceloc`.
 - The file path is no longer split into `base_dir` and `filename`, it's
   now just `full_path`.
 - There's only `function.symbol`, no more `function.function_name`.
   I've removed a FIXME comment about this.
 - One of the tests notices that a double slash is turned into a single
   slash. That's fine, I adjusted the test.

The more interesting part of this patch is the emitting of inline frames.
They are put into a per-frame `inlines` property, ordered from
callee-most to caller-most.
The test which I added for this actually found a bug in symbolic's
symcache code, which is now fixed in 9.1.1.