bug 1796579: merge /api/uploads/ views into one
This merges all the /api/uploads/* views back into a single view.

This changes the way the "total" is handled. If there are no filters,
then it doesn't calculate a total and instead uses 1,000,000 which is a
"magic big number" that the frontend understands to be "I didn't
calculate the total". If there are filters, then it'll calculate the
total.