bug 1796633: merge /api/uploads/files/ views into one
This merges all the /api/uploads/files/* back into a single view.

This changes how "total" is handled. If there are no filters, then it
doesn't calculate a total and instead uses 1,000,000 which is a magic
number the frontend understands as "many". If there are filters, then
it'll calculate the total.