bug 1800479: fix ParseSymFileError to be more helpful
Previously, ParseSymFileError didn't include an exception message so
debugging these was tricky because you knew there was an error, but it
didn't include enough detail as to why.

This also fixes the handling cade to use logger.exception rather than
logger.error so we get a traceback which helps debug issues.