bug 1793984: reorder frames to symbolicate
This reworks symbolication to reorder all the frames to symbolicate by
module so that as we're symbolicating frames, we only need one module in
memory at a time.

This also removes the need for using a temporary directory to store a
symbol file that was just downloaded in order to parse it with symbolic.
Now we parse it straight from bytes.