bug 1817902: add orphaned files removal
When a gunicorn worker is handling a symbol upload request and it takes
too long, gunicorn will kill the worker off orphaning the workspace the
worker was using. This causes the disk to fill up and then the node to
fail.

This fixes that by adding a "disk manager" process that wakes up
periodically to remove files that have been around too long and have
likely been orphaned.