bug 1817902: redo remove_orphaned_files output; fix logging issues
This redoes the output for remove_orphaned_files so it's logged rather
than written to stdout/stderr. In order to test that, I had to fix how
logging was configured in tecken so that it all went through the
root--that way I can use caplog in the tests.

While doing that, I fixed honcho to not prepend prefixes to log lines
and I added the "processname" (as set by the run scripts) to the log
metadata.