bug 1832823: fix list-firefox-symbols-zips
This fixes the list-firefox-symbols-zips script to handle changes they
made in firefox ci where they purged a bunch of symbols zip files, but
didn't update the expiry in the artifact records.

This skips tasks that are more than 3 months old (1 year - 9 months) and
adds skipping for artifacts that have expired as well.