bug 1843356: fix check_url_head metrics
Tecken emits metrics for the cache of whether things exist in the
storage buckets. Prior to the removal of the private symbols bucket
support, there were two functions for this: one that used boto and one
that used HTTP. When I removed the private symbols bucket support, I
should have removed the boto one since we don't use it anymore and I
should have shifted the metrics names so we maintained the cache
metric keys.

This fixes that.