bug 1847937: fix upload file view for files in incomplete uploads
If an upload is interrupted, Tecken doesn't end up with a correct count
of the number of files in the upload. It only knows about the files it
has processed so far or was in the process of processing. This fixes
some language in the UI around that.

If an upload is interrupted, when you go to look at an upload file view
for a file in the upload, it displays some stats about the upload
including a DisplayIncompleteRatio section which required data that
wasn't in the API response. Because Tecken doesn't know how many files
were in the upload, I decided to just nix the DisplayIncompleteRatio
section altogether because the information it was trying to render
wasn't meaningful.