bug 1746940: implement code_file/code_id lookup for download API
When the Socorro processor stackwalker requests sym files with bogus
debug id 000000000000000000000000000000000 Tecken would return a 404
because it had no idea what sym file was being requested.

This changes the download API so that if there are valid code_file and
code_id querystring params, it'll look up the FileUpload record and
return the correct debug_id. Then it'll issue an HTTP 302 with the
correct download API url.

This keeps track of how often it does a successful code_file/code_id
lookup with:

* `tecken.download_symbol_code_id_lookup`