bug 1746940: fix FileUpload admin; add fake data generator
This fixes the FileUpload admin which wasn't usable in prod before
because it'd create a drop-down for upload which had ... 300k items in
it. We'll never want to change that field value, so now it's readonly.

This adds a fake data generator that's kind of terrible, but good enough
for now and something we can improve over time as needs arise.