bug 1746940: rework download API to support /code_file/code_id/sym_file
This reworks the download API to support /code_file/code_id/sym_file
urls. If it sees this, it'll look up code_file/code_id in the fileupload
records. If it finds one, it'll return an HTTP 302 with the correct
debug_filename and debug_id.

This works for both HEAD and GET requests.

This updates the download API documentation as well.