bug 1746940: move info lookup to syminfo API endpoint
This moves the codefile/codeid lookup to its own API endpoint and also
adds support for looking up by debugfile/debugid. This is generally
useful since it provides additional symbol file information. Further, it
makes it easier to test the database query and improve it before we put
it back in the heavily used download API.