bug 1746940: implement codeinfo lookup in download API
This re-implements the codeinfo lookup in the download API which kicks
in if a symbol file wasn't found and if the debugfilename/debugid look
like a possible codefile/codeid.

There's one scenario which is a little weird:

1. client uses non-try download url with code info
2. view performs a code info lookup and gets a try-symbol result
3. view returns a non-try download url redirect with try-symbol debug
   info details
4. client requests redirect and gets a 404

The eventual outcome (404) is correct, but it goes through an extra
redirect. I decided not to account for that because it involves storing
storage backend specific information in a way that's hard to reason
about and makes it harder to rewrite the storage api later.

If it turns out this was a bad choice, we can rethink it later.