bug-1853962: change UPLOAD_TEMPDIR does not exist into non-error
When Tecken starts up, the UPLOAD_TEMPDIR directory may not exist. It
(or its parents) may not get created until an upload API request is
handled.

We could rework everything so it gets created earlier. Alternatively, we
could fix remove_orphaned_files to note that the directory doesn't exist
and then move on. I opted for the latter for now.