bug-1853962: rework sentry-wrap output handling
The Python logging StreamHandler defaults to printing logging output to
stderr. This is a problem when used with sentry-wrap because if the
process exited with a 0 exit code (success!), then all the output would
be in stderr and ignored.

This reworks the sentry-wrap output handling so that it multiplexes
stdout and stderr and then prints it regardless of whether the
subprocess succeeds or fails.

This also truncates the output to the last 5,000 characters before
sending it as a Sentry event just in case the output is monstrous.