bug-1867587: add slick.sh-dist
This adds slick.sh-dist which makes it easier to recreate Tecken local
development environments with a superuser account and an API token. It
adds a make rule for building it.

In order to do that, it redoes how we copy dist files and updates .env
copying to use the same stuff.