bug-1847235: remove markus configuration in gunicornhooks
worker_abort is run in the sigabrt handling in the gunicorn worker. That
process has markus and logging configured already, so we don't need to
additionally configure it in this file.

Further, this adds comments clarify the context things are run in and
the rules for adding things to the gunicornhooks.py file.