bug-1866796: ensure system tests upload and download from the right bucket
Because:
* Upload and download system tests wrongly assume all symbols files are in the regular bucket.
* The first download system test (which based on the current `sym_files_to_download.csv`, corresponds to a try file) was failing when running system tests in prod.

This commit:
* Adds additional columns to the CSV generated for the download tests to indicate which bucket the file is in (and, while we're at it, which platform the file is from).
* Ensures the download system tests check the bucket and modifies the download request accordingly.
* Ensures the upload system tests check the bucket and modifies the upload request accordingly.

Note: I made some manual changes to sym_files_to_download.csv, as it seems like the `./test_env.sh` shell script was terminating the Python process for `download-sym-files.py` on `request.get` for files on the order of 150MB. I replaced such large files with smaller files. This is some kind of OOM process killer AFAICT. Will need to address in a later commit.