bug-1881632: update circleci remote docker image
Previously, we specified an image that's now deprecated. This removes
the version specification which will cause it to use the default which
is currently docker 24. default will change over time, but we'll know if
CI fails pretty quick and we're switching to GCP soon with a completely
different deploy pipeline anyhow.