bug-1906597: Remove all uses of Upload.bucket_endpoint_url.
The bucket_endpoint_url field is always set to NULL for our prod and
stage instances. It only has non-NULL values in the devstack. It would
be possible to set the field to the actual HTTPS endpoint for the
bucket, but I don't think we need the field, so I removed all uses
instead. We could remove the field in a later data migration.