bug-1827507: Fix GCS emulator settings.
The GCS emulator needs special settings to allow public downloads and
retriable uploads.

Public downloads are required so we can redirect to files in the GCS
emulator from the download API. They are enabled by the `-public-host`
command-Line option.

Retriable uploads are used by the `Blob.upload_from_file()` method. They
return a redirect in the initial request, which requires the server to
know its external URL, which can be set using `-external-url`.