bug-1910613: update guidance on 429s and symbols zip archive sizes
This updates the guidance on retrying to wait 15 seconds and try again.

This also updates the discussion around which HTTP POST method to use in
relation to the symbol zip archive size. We now recommend this new model
where we encourage people to have symbol zip archive sizes < 2gb and
they should split the archive up across multiple HTTP POSTs.